requirements.txt
streamlit>=1.36.0
langchain>=0.2.12
langchain-core>=0.2.30
langchain-community>=0.2.10
langchain-openai>=0.1.7
langchain-text-splitters>=0.2.2
chromadb>=0.5.4
pypdf>=4.2.0
python-dotenv>=1.0.1
tiktoken>=0.7.0
pydantic>=2.7.0
